fix: show hidden scrollbars after unmounting
Configuring Fullpage with the `hideScrollBars` option set to `true`
naturally hides the scroll bar on the page the component is first used.
However, a (possibly) unintended side-effect is that the scroll bar
remains hidden even after the `<Fullpage>` component (and its
ancestors) are unmounted.

The lines introduced by this commit ensure the scroll bar is shown
again during the execution of the `componentWillUnmount()` lifecycle
method, providing the `hideScrollBars` option is set to `true`.

This results in Fullpage properly cleaning up after itself when it
unmounts from the tree, meaning that one can switch between components
(e.g. pages) that use and don't use Fullpage assisted scrolling without
any side-effects.